+++ cmd/user/dal/db/video.go
@@ -22,13 +22,13 @@ import (
 //		}
 type Video struct {
 	gorm.Model
-	ID            int       `gorm:"column:id;primary_key;AUTO_INCREMENT"`
-	AuthorID      int       `gorm:"column:author_id;NOT NULL"`
+	ID            int64     `gorm:"column:id;primary_key;AUTO_INCREMENT"`
+	AuthorID      int64     `gorm:"column:author_id;NOT NULL"`
 	PublishTime   time.Time `gorm:"column:publish_time;default:CURRENT_TIMESTAMP;NOT NULL"`
 	FilePath      string    `gorm:"column:file_path;NOT NULL"`
 	CoverPath     string    `gorm:"column:cover_path;NOT NULL"`
-	FavoriteCount int       `gorm:"column:favorite_count;default:0"`
-	CommentCount  int       `gorm:"column:comment_count;default:0"`
+	FavoriteCount int64     `gorm:"column:favorite_count;default:0"`
+	CommentCount  int64     `gorm:"column:comment_count;default:0"`
 	Title         string    `gorm:"column:title;NOT NULL"`
 }
